Add location selection by moving map, pin title input, photo picker and map UI improvements

+++ Pin+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Pin+CoreDataProperties.swift
 //  journeypin
 //
-//  Created by Al Sneed on 03/06/2025.
+//  Created by maria on 14/06/2025.
 //
 //
 
@@ -16,8 +16,8 @@ extension Pin {
         return NSFetchRequest<Pin>(entityName: "Pin")
     }
 
-    @NSManaged public var lat: Float
-    @NSManaged public var lon: Float
+    @NSManaged public var lat: Double
+    @NSManaged public var lon: Double
     @NSManaged public var name: String?
     @NSManaged public var pinphoto: NSSet?
     @NSManaged public var pintrip: Trip?
